api/requirements.txt
asgiref==3.4.1
CacheControl==0.12.6
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
cryptography==35.0.0
defusedxml==0.7.1
Django==3.1.14
django-allauth==0.44.0
django-cors-headers==3.10.0
django-filter==21.1
django-phonenumber-field==5.2.0
django-storages==1.12
django-twilio==0.13.1.post0
djangorestframework==3.12.4
django-rest-auth==0.9.5
#djangorestframework-simplejwt==4.6.0
#firebase-admin==5.0.3
google-api-core==2.1.0
google-api-python-client==2.24.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
gunicorn==20.1.0
httplib2==0.20.1
idna==3.2
Markdown==3.3.4
msgpack==1.0.2
oauthlib==3.1.1
packaging==21.0
phonenumbers==8.12.34
Pillow==8.3.2
proto-plus==1.19.4
protobuf==3.18.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-http-client==3.3.3
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sendgrid==6.8.2
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.1
twilio==6.63.2
uritemplate==3.0.1
urllib3==1.26.7
